# "Canary" changes for user column inputs