--- a/nemosis/data_fetch_methods.py
+++ b/nemosis/data_fetch_methods.py
@@ -83,7 +83,7 @@ def dynamic_data_compiler(start_time, end_time, table_name, raw_data_location,
         print(f'Returning {table_name}. Check warnings for failed downloads '
               + 'or failure to load data')
         return all_data
-    else:
+    elif data_merge:
         print(f'Compiling data for table {table_name} FAILED.')