… nemosis/processing_info_maps.py