# Rationale
As of #15, there were some issues with user input handling:
## Issues
1. If a data file existed in cache, NEMOSIS went into a try-except block that tried to read the file with user-provided `select_columns`. For feather and parquet, the read will fail if the columns are not in the file, and the exception handler would just read the file without any filter applied. This is not expected behaviour when the user provides a filter.
2. `select_columns` would remove any columns present in the data that were not included in `select_columns`. If a column does not exist in the data, it is "ignored" - this provides no user feedback that the provided column is not in the data.
3. If a datetime filter is provided, but `select_columns` does not include the appropriate datetime column, NEMOSIS will error out. Expected behaviour would be that date filtering is applied but then only selected columns are returned. (i.e. the user would not be aware these are done sequentially in the code).
4. `filter_cols` has to be a part of the `select_columns`, otherwise NEMOSIS will error out with no obvious error message.

# Changes
- `_validate_select_columns` added, which checks whether `select_columns` exist in the data. Also provides user feedback letting the user know that some filters were applied, some were not (i.e. will return partially column-filtered DataFrame based on valid cols). If `select_columns` are all not in the data, NEMOSIS will error out and provide useful feedback (addresses issues 1 and 2).
- The previous function also returns datetime cols so datetime filtering can occur. If the datetime cols are not requested by the user, they will be dropped prior to returning data. To do this, a new `date_col` mapping was added in `processing_info_maps.py` (addresses issue 3).
- `filter_col` validation carried out prior to running fetch loop (are `filter_cols` a subset of `select_columns` after the latter is preprocessed?) (addresses issue 4).
- Tests modified because of new function and additional function args in key functions (e.g. fetch loop)

# Outstanding changes
- Enhancing user logging and feedback - #17

## Examples of feeedback
**Successful download + load**
![image](https://user-images.githubusercontent.com/29934785/119427231-07269580-bd4e-11eb-98ca-759e5c143322.png)
**Successful load**
![image](https://user-images.githubusercontent.com/29934785/119427426-5967b680-bd4e-11eb-8539-289ae96e99fb.png)
**Successful `select_columns` with date filtering but date column not returned**
![image](https://user-images.githubusercontent.com/29934785/119427506-7d2afc80-bd4e-11eb-9b6e-bab5b6210148.png)
**As above, but with date col returned**
![image](https://user-images.githubusercontent.com/29934785/119429323-fa0ba580-bd51-11eb-845d-71223d503cd7.png)
**Succesful `select_columns` with date filtering and one invalid col provided**
![image](https://user-images.githubusercontent.com/29934785/119427587-9cc22500-bd4e-11eb-9d29-e7cf1a5dfbe4.png)
**Partially successful load (one file not downloaded and hence not loaded)**
![image](https://user-images.githubusercontent.com/29934785/119428236-e0695e80-bd4f-11eb-9c66-e8b77115d625.png)
**Unsuccesful load as `select_columns` invalid**
![image](https://user-images.githubusercontent.com/29934785/119427641-b6636c80-bd4e-11eb-9571-aec68c1f63fc.png)
**Unsuccesful load as `select_columns` invalid (but `filter_cols` in `select_columns`)**
![image](https://user-images.githubusercontent.com/29934785/119428017-7c469a80-bd4f-11eb-8aad-a789e67a8567.png)
**Unsuccesful load as `filter_cols` not in `select_columns`**
![image](https://user-images.githubusercontent.com/29934785/119428071-92ecf180-bd4f-11eb-9f47-d4208dc0953b.png)
![image](https://user-images.githubusercontent.com/29934785/119428108-a26c3a80-bd4f-11eb-8821-0d724d01603f.png)
**Failure to load data (last line only returned if `data_merge` is True, otherwise download warnings + file creation failures printed)**
![image](https://user-images.githubusercontent.com/29934785/119428874-0d6a4100-bd51-11eb-9aae-20b10719b50c.png)

# Performance
![image](https://user-images.githubusercontent.com/29934785/119431150-689e3280-bd55-11eb-975a-66de730bf162.png)
- New changes: 3.54 s +/- 102 ms per loop, 7 loops
- Old (as of 85e9ebe): 2.04s +/- 59.7ms, 7 loops
- 70% slower, due to increased number of `if` statements
  - Is user feedback prioritised over speed?

# Testing
- Passed all (modified) tests: 
[Test Report.pdf](https://github.com/UNSW-CEEM/NEMOSIS/files/6536073/Test.Report.pdf)